Updated project and package name
