+++ conga-console/build.gradle
@@ -7,7 +7,7 @@ plugins {
 }
 
 dependencies {
-    implementation project(':api')
+    implementation project(':conga-core')
 }
 
 application {
