+++ conga-core/src/main/java/ai/conga/core/domain/Copy.java
@@ -1,4 +1,4 @@
-package conga.ai.api;
+package ai.conga.core.domain;
 
 /**
  *
