+++ settings.gradle
@@ -1,2 +1,2 @@
 rootProject.name = 'conga-ai'
-include('app', 'api')
+include('conga-core', 'conga-console')
